Problemas com BigDecimal em Product resolvidos

+++ src/main/java/learn/spring/api/dtos/ProductRecordDTO.java
@@ -1,10 +1,8 @@
 package learn.spring.api.dtos;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record ProductRecordDTO(@NotBlank String name, @NotNull BigDecimal value) {
+public record ProductRecordDTO(@NotBlank String name, @NotNull Double value) {
     //Por padrão, seus atributos já são final e private
 }
